Rework a bunch of imports; consolidate types
- Move types into `@music/data/types`
- Fix all the relative imports to be `@music/`
- Fix `@components/` paths
