+++ next-js/app/music/bucket-list/page.tsx
@@ -1,14 +1,14 @@
 import { Metadata } from "next";
 import { routes } from "@music/lib/routes";
 import { IndexPage } from "@music/components/IndexPage";
-import { formatWorkTitle, formatComposerName } from "../lib/helpers";
+import { formatWorkTitle, formatComposerName } from "@music/lib/helpers";
 import {
   getOrderedBucketList,
   getWorkByTitle,
 } from "@music/data/queries/works";
 import { getComposerByTitle } from "@music/data/queries/composers";
 import { getConcertsByWork } from "@music/data/queries/concerts";
-import { Work } from "../lib/types";
+import { Work } from "@music/data/types";
 
 export const metadata: Metadata = {
   title: "Bucket List",
