+++ next-js/app/music/components/ConcertInfo.tsx
@@ -7,11 +7,11 @@ import {
   isUpcoming,
 } from "@music/lib/helpers";
 import { routes } from "@music/lib/routes";
-import { Concert } from "@music/lib/types";
+import { Concert } from "@music/data/types";
 import { getLocationFromCoordinates } from "@music/lib/location";
 import { getGroupByTitle } from "@music/data/queries/groups";
 import { getVenueByTitle } from "@music/data/queries/venues";
-import { AttendActions } from "./AttendActions";
+import { AttendActions } from "@music/components/AttendActions";
 
 interface ConcertInfoProps {
   concert: Concert;
