+++ next-js/app/music/conductors/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from "next";
 import { notFound } from "next/navigation";
-import { PageProps } from "@music/lib/types";
+import { PageProps } from "@music/data/types";
 import { ConcertListItem } from "@music/components/ConcertListItem";
 import { getDateForSorting } from "@music/lib/helpers";
 import { PageTitle } from "@music/components/PageTitle";
