+++ next-js/app/music/data/queries/works.ts
@@ -1,5 +1,5 @@
-import database from "../database";
-import { Work } from "@music/lib/types";
+import database from "@music/data/database";
+import { Work } from "@music/data/types";
 
 export function getWorks() {
   return database.work;
